services/notification-service/requirements.txt
fastapi
uvicorn[standard]
nats-py
resend
python-dotenv
